feat: add hideModal function for modal management and refactor modal hiding in login and register flows
